refactor: correct file's name for kebab-case and alter to absolute importation
