+++ src/domain/entities/customer/errors/invalid-name.ts
@@ -1,5 +1,5 @@
 export class InvalidName extends Error {
   constructor() {
-    super('Name is invalid');
+    super('O nome deve ter pelo menos 3 caracteres e apenas letras');
   }
 }
