+++ src/domain/entities/customer/types/customer-entity-props.ts
