+++ src/domain/entities/customer/types/customer-props-errors.ts
@@ -1,7 +1,7 @@
+import { CustomerEntity } from '@/domain/entities/customer/customer-entity';
 import { InvalidName } from '@/domain/entities/customer/errors';
 import { FieldIsRequired, InvalidField } from '@/domain/shared/errors';
 import { Either } from '@/shared/either';
-import { CustomerEntity } from '../customer-entity';
 
 export type ErrorsCustomerProps = FieldIsRequired | InvalidField | InvalidName;
