+++ src/domain/entities/customer/types/customer-props.ts
