+++ src/domain/entities/customer/types/index.ts
@@ -1,3 +1,3 @@
-export * from './customerEntityProps';
-export * from './customerProps';
-export * from './customerPropsErrors';
+export * from './customer-entity-props';
+export * from './customer-props';
+export * from './customer-props-errors';
